requirements.txt
APScheduler==3.6.3
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
blis==0.7.7
cachetools==4.2.2
catalogue==2.0.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cycler==0.11.0
cymem==2.0.6
decorator==5.1.1
emoji==1.7.0
executing==0.8.3
fonttools==4.31.2
idna==3.3
ipython==8.2.0
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.2
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
murmurhash==1.0.6
nltk==3.7
numpy==1.22.3
packaging==21.3
pandas==1.4.1
parso==0.8.3
pathy==0.6.1
pickleshare==0.7.5
Pillow==9.1.0
preshed==3.0.6
prompt-toolkit==3.0.29
pure-eval==0.2.2
pydantic==1.8.2
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-telegram-bot==13.10
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.15
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
smart-open==5.2.1
spacy==3.2.3
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.2
stack-data==0.2.0
thinc==8.0.15
threadpoolctl==3.1.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typer==0.4.1
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
wasabi==0.9.1
wcwidth==0.2.5
